[feat] hook수정
