+++ src/views/Mypage/hooks/useUserUpdate.ts
@@ -7,7 +7,7 @@ export const useUserUpdate = (
     newPassword: string,
     passwordCheck: string,
     nickname: string,
-    handleModalClose: () => void,
+    closeModal: () => void,
     refetchUserInfo: () => void,
 ) => {
     // const token = useAuthToken();
@@ -37,7 +37,7 @@ export const useUserUpdate = (
 
             //서버에서 받은 새로운 정보 상태 업데이트
             refetchUserInfo();
-            handleModalClose();
+            closeModal();
         } catch (error) {
             console.error('회원정보 수정 실패', error);
             alert('회원정보 수정에 실패했습니다.');
